Move isAllOf/isAnyOf to one file, and de-dupe some code.
