+++ src/matchers.ts
@@ -1,0 +1,58 @@
+import {
+  ActionFromMatcher,
+  hasMatchFunction,
+  Matcher,
+  UnionToIntersection
+} from './tsHelpers'
+
+/** @public */
+export type ActionMatchingAnyOf<
+  Matchers extends [Matcher<any>, ...Matcher<any>[]]
+> = ActionFromMatcher<Matchers[number]>
+
+/** @public */
+export type ActionMatchingAllOf<
+  Matchers extends [Matcher<any>, ...Matcher<any>[]]
+> = UnionToIntersection<ActionMatchingAnyOf<Matchers>>
+
+const matches = (matcher: Matcher<any>, action: any) => {
+  if (hasMatchFunction(matcher)) {
+    return matcher.match(action)
+  } else {
+    return matcher(action)
+  }
+}
+
+/**
+ * A higher-order function that returns a function that may be used to check
+ * whether an action matches any one of the supplied type guards or action
+ * creators.
+ *
+ * @param matchers The type guards or action creators to match against.
+ *
+ * @public
+ */
+export function isAnyOf<Matchers extends [Matcher<any>, ...Matcher<any>[]]>(
+  ...matchers: Matchers
+) {
+  return (action: any): action is ActionMatchingAnyOf<Matchers> => {
+    return matchers.some(matcher => matches(matcher, action))
+  }
+}
+
+/**
+ * A higher-order function that returns a function that may be used to check
+ * whether an action matches all of the supplied type guards or action
+ * creators.
+ *
+ * @param matchers The type guards or action creators to match against.
+ *
+ * @public
+ */
+export function isAllOf<Matchers extends [Matcher<any>, ...Matcher<any>[]]>(
+  ...matchers: Matchers
+) {
+  return (action: any): action is ActionMatchingAllOf<Matchers> => {
+    return matchers.every(matcher => matches(matcher, action))
+  }
+}
